create endpoint for range of dates

+++ backend/api/apis/prow/v1alpha1/metrics.go
@@ -86,5 +86,5 @@ type UnknowFailuresImpact struct {
 	Total int `json:"total"`
 
 	// The percentage of the total job impacted
-	Percentage int `json:"percentage"`
+	Percentage float64 `json:"percentage"`
 }
